move meta structs around
