+++ midix/src/file/meta/message.rs
@@ -4,6 +4,61 @@ use super::{KeySignatureRef, TempoRef, TimeSignatureRef};
 
 /// A "meta message", as defined by the SMF spec.
 /// These events carry metadata about the track, such as tempo, time signature, copyright, etc...
+#[derive(Clone, PartialEq, Eq, Debug, Hash)]
+pub enum MetaMessage {
+    /// For `Format::Sequential` MIDI file types, `TrackNumber` can be empty, and defaults to
+    /// the track index.
+    TrackNumber(Vec<u8>),
+    /// Arbitrary text associated to an instant.
+    Text(Vec<u8>),
+    /// A copyright notice.
+    Copyright(Vec<u8>),
+    /// Information about the name of the track.
+    TrackName(Vec<u8>),
+    /*/// Information about the name of the current instrument.
+    InstrumentName(&'a [u8]),
+    /// Arbitrary lyric information associated to an instant.
+    Lyric(&'a [u8]),
+    /// Arbitrary marker text associated to an instant.
+    Marker(&'a [u8]),
+    /// Arbitrary cue point text associated to an instant.
+    CuePoint(&'a [u8]),
+    /// Information about the name of the current program.
+    ProgramName(&'a [u8]),
+    /// Name of the device that this file was intended to be played with.
+    DeviceName(&'a [u8]),
+    /// Number of the MIDI channel that this file was intended to be played with.
+    MidiChannel(Channel),
+    /// Number of the MIDI port that this file was intended to be played with.
+    MidiPort(u8),
+    /// Obligatory at track end.
+    EndOfTrack,
+    /// Amount of microseconds per beat (quarter note).
+    ///
+    /// Usually appears at the beginning of a track, before any midi events are sent, but there
+    /// are no guarantees.
+    Tempo(TempoRef<'a>),
+    /// The MIDI SMPTE offset meta message specifies an offset for the starting point of a MIDI
+    /// track from the start of a sequence in terms of SMPTE time (hours:minutes:seconds:frames:subframes).
+    ///
+    /// [Reference](https://www.recordingblogs.com/wiki/midi-smpte-offset-meta-message)
+    SmpteOffset(&'a [u8]),
+    /// In order of the MIDI specification, numerator, denominator, MIDI clocks per click, 32nd
+    /// notes per quarter
+    TimeSignature(TimeSignatureRef<'a>),
+    KeySignature(KeySignatureRef<'a>),
+    /// Arbitrary data intended for the sequencer.
+    /// This data is never sent to a device.
+    SequencerSpecific(&'a [u8]),
+    /// An unknown or malformed meta-message.
+    ///
+    /// The first `u8` is the raw meta-message identifier byte.
+    /// The slice is the actual payload of the meta-message.
+    Unknown(&'a u8, &'a [u8]),*/
+}
+
+/// A "meta message", as defined by the SMF spec.
+/// These events carry metadata about the track, such as tempo, time signature, copyright, etc...
 #[derive(Copy, Clone, PartialEq, Eq, Debug, Hash)]
 pub enum MetaMessageRef<'a> {
     /// For `Format::Sequential` MIDI file types, `TrackNumber` can be empty, and defaults to
