+++ midix/src/file/track/message.rs
@@ -1,6 +1,15 @@
 use crate::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
+pub enum MidiTrackMessage {
+    ChannelVoice(ChannelVoiceMessage),
+    SystemExclusive(SystemExclusive),
+    /// A meta-message, giving extra information for correct playback, like tempo, song name,
+    /// lyrics, etc...
+    Meta(MetaMessage),
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Hash)]
 pub enum MidiTrackMessageRef<'a> {
     ChannelVoice(ChannelVoiceMessage),
     SystemExclusive(SystemExclusiveRef<'a>),
